Fix itemrack set matching in tooltips
Use exact IDs for items in bags/inventory, and base IDs for links.

This fixes the problem where you might have two of the same item
but with different runes applied.  Before this fix, ItemRack
would show both items as belonging to both sets, which is only true
if the exact item from the set wasn't found at time of equip.

Now, we'll only show set information in tooltips for items in bags/
inventory if that exact item is part of a set.  Links from other
players will still show sets you have that include the same base ID.